add to Cart, delete to Cart

+++ src/main/java/com/hcmute/mobilestore/models/Product.java
@@ -77,7 +77,7 @@ public class Product {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(float price) {
         this.price = price;
     }
 }
